fix(preview): Honeycomb int32 cast + ResizeObserver-based canvas scaling
Two issues from manual testing of the previous push:

## 1. Honeycomb failed to initialize in the browser

`heart-matrix-controller/patterns/honeycomb.py` raises
`TypeError: Cannot cast array data from dtype('int64') to dtype('int32')`
in Pyodide's bundled numpy because `np.choose(i, ...)` in numpy 2.x
requires int32 indices. Cast `np.floor(h).astype(np.int64)` to
np.int32 (the comment notes this is for the browser; the Pi is on
numpy 1.x and doesn't hit the cast error, but int32 is valid in both).

Result: the browser cycle now includes all four patterns (Fireworks,
Flame, NightSky, Honeycomb). No code change on the Pi side; the cast
is a no-op for the existing Pi's numpy version.

## 2. Canvas didn't scale with the viewport on resize

Two bugs in `static/preview/preview.js`:

(a) The width math subtracted 64px (assumed 32px on each side), but
    the actual padding chain is `p-12` (48px on each side) on the
    `bg-white` card + `p-4` (16px on each side) on the dark div
    = 128px total. On a 600px-wide viewport, the math gave 512px
    but the card could only hold ~440px, so the canvas overflowed
    the card. The bg-white card's clientWidth gives the actual
    available width; fall back to `window.innerWidth - 160` if
    the card selector no longer matches.

(b) `window.addEventListener('resize', ...)` doesn't fire for all
    viewport changes (URL bar collapse on mobile, devtools docked
    to the side on some browsers). Switched the primary path to a
    `ResizeObserver` on the bg-white card — fires for every
    layout change, uses the card's actual clientWidth, and removes
    the circular reference of observing the canvas (whose parent
    sizes to the canvas). Kept the `window.resize` fallback for
    browsers without `ResizeObserver` (none in our support matrix,
    but defensive).

Tests: 101/101 pass (added 2 new tests for the ResizeObserver path
and the card.clientWidth computation). 99 previous tests unchanged.

Refs #8.

+++ heart-matrix-controller/patterns/honeycomb.py
@@ -37,7 +37,11 @@ def _triangle(v):
 def _hsv_to_rgb(h, s, v):
     """Vectorized HSV->RGB. h,v are arrays in 0..1, s scalar; returns (...,3)."""
     h = (h % 1.0) * 6.0
-    i = np.floor(h).astype(np.int64)
+    # np.choose in numpy 2.x (shipped in Pyodide) requires int32 indices;
+    # int64 raises "Cannot cast array data from dtype('int64') to
+    # dtype('int32')". Cast here so the pattern works in both the
+    # browser (Pyodide) and on the Pi.
+    i = np.floor(h).astype(np.int32)
     f = h - i
     p = v * (1.0 - s)
     q = v * (1.0 - f * s)
